Move energy-related logic to a new EnergyManager.
