+++ src/main/java/emu/grasscutter/game/player/Player.java
@@ -24,6 +24,7 @@ import emu.grasscutter.game.inventory.Inventory;
 import emu.grasscutter.game.mail.Mail;
 import emu.grasscutter.game.mail.MailHandler;
 import emu.grasscutter.game.managers.StaminaManager.StaminaManager;
+import emu.grasscutter.game.managers.EnergyManager;
 import emu.grasscutter.game.managers.SotSManager;
 import emu.grasscutter.game.props.ActionReason;
 import emu.grasscutter.game.props.EntityType;
@@ -144,6 +145,7 @@ public class Player {
 
 	@Transient private MapMarksManager mapMarksManager;
 	@Transient private StaminaManager staminaManager;
+	@Transient private EnergyManager energyManager;
 
 	private long springLastUsed;
 	private HashMap<String, MapMark> mapMarks;
@@ -194,6 +196,7 @@ public class Player {
 		this.mapMarksManager = new MapMarksManager(this);
 		this.staminaManager = new StaminaManager(this);
 		this.sotsManager = new SotSManager(this);
+		this.energyManager = new EnergyManager(this);
 	}
 
 	// On player creation
@@ -222,6 +225,7 @@ public class Player {
 		this.mapMarksManager = new MapMarksManager(this);
 		this.staminaManager = new StaminaManager(this);
 		this.sotsManager = new SotSManager(this);
+		this.energyManager = new EnergyManager(this);
 	}
 
 	public int getUid() {
@@ -1087,6 +1091,10 @@ public class Player {
 	public StaminaManager getStaminaManager() { return staminaManager; }
 
 	public SotSManager getSotSManager() { return sotsManager; }
+
+	public EnergyManager getEnergyManager() {
+		return this.energyManager;
+	}
 
 	public AbilityManager getAbilityManager() {
 		return abilityManager;
